Restyle password reset request form
