+++ components/authentication/ResetPasswordForm.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { CardWrapper } from "../CardWrapper";
 import { PasswordResetSchema } from "@/schemas/AuthValidation";
 import { useForm } from "react-hook-form";
 import { useState, useTransition } from "react";
@@ -9,15 +8,16 @@ import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "
 import * as z from 'zod';
 import { Input } from "../ui/input";
 import { Button } from "../ui/button";
-import { FormError } from "../FormError";
-import { FormSuccess } from "../FormSuccess";
 import { PasswordResetResult, resetPassword } from "@/actions/PasswordReset";
+import React from "react";
+import { AlertType, AuthAlert } from "./Alerts";
+import Link from "next/link";
 
 export const ResetPasswordForm = () => {
     const [isPending, startTransition] = useTransition();
     const [error, setError] = useState<string | undefined>("");
     const [success, setSuccess] = useState<string | undefined>("");
-  
+
     const form = useForm<z.infer<typeof PasswordResetSchema>>({
         resolver: zodResolver(PasswordResetSchema),
         defaultValues: {
@@ -26,47 +26,60 @@ export const ResetPasswordForm = () => {
     });
 
     const onSubmit = (values: z.infer<typeof PasswordResetSchema>) => {
-            setSuccess("");
-            setError("");
+        setSuccess("");
+        setError("");
 
-            startTransition(() => {
-                resetPassword(values)
-                    .then((data: PasswordResetResult | undefined) => {
-                        setError(data?.error);
-                        setSuccess(data?.success);
-                    })
-            });
+        startTransition(() => {
+            resetPassword(values)
+                .then((data: PasswordResetResult | undefined) => {
+                    setError(data?.error);
+                    setSuccess(data?.success);
+                })
+        });
     };
 
     return (
-        <CardWrapper headerName="Reset password" backButtonText="Back to Login" backButtonHref="/login">
+        <>
             <Form {...form}>
-                <form
-                    className="space-y-6"
-                    onSubmit={form.handleSubmit(onSubmit)}>
-                    <section className="space-y-4">
-                        <FormField control={form.control} name="email" render={({ field }) => (
-                            <FormItem>
-                                <FormLabel>Email</FormLabel>
-                                <FormControl>
-                                    <Input
-                                        {...field}
-                                        disabled={isPending}
-                                        placeholder="Enter your email address"
-                                        type="email"
-                                    />
-                                </FormControl>
-                                <FormMessage></FormMessage>
-                            </FormItem>
-                        )} />
-                    </section>
-                    <FormSuccess message={success} hasCloseButton={true} />
-                    <FormError message={error} hasCloseButton={true} />
-                    <Button type="submit" className="w-full" disabled={isPending}>
-                        Send password reset email
-                    </Button>
+                <form className="p-6 md:p-8" onSubmit={form.handleSubmit(onSubmit)}>
+                    <div className="flex flex-col gap-6">
+                        <div className="flex flex-col items-center text-center">
+                            <h1 className="text-2xl font-bold">Welcome back!</h1>
+                            <p className="text-balance text-muted-foreground">
+                                Request a password reset
+                            </p>
+                        </div>
+                        <div className="grid gap-2">
+                            <React.Fragment>
+                                <FormField control={form.control} name="email" render={({ field }) => (
+                                    <FormItem>
+                                        <FormLabel>Email address</FormLabel>
+                                        <FormControl>
+                                            <Input
+                                                {...field}
+                                                disabled={isPending}
+                                                placeholder="Enter your email address"
+                                                type="email"
+                                            />
+                                        </FormControl>
+                                        <FormMessage></FormMessage>
+                                    </FormItem>
+                                )} />
+                            </React.Fragment>
+                        </div>
+                        <AuthAlert message={success} hasCloseButton={true} type={AlertType.SUCCESS} />
+                        <AuthAlert message={error} hasCloseButton={true} type={AlertType.ERROR} />
+                        <Button type="submit" className="w-[60%] text-center m-auto font-extrabold">
+                            Request password reset
+                        </Button>
+                        <div className="text-center text-sm">
+                            <Link href="/login" className="underline underline-offset-4">
+                                Back to login
+                            </Link>
+                        </div>
+                    </div>
                 </form>
             </Form>
-        </CardWrapper>
+        </>
     );
 };
